# perf: optimize game map rendering