--- a/src/simulat/core/surfaces/game_map/game_map.py
+++ b/src/simulat/core/surfaces/game_map/game_map.py
@@ -78,11 +78,41 @@ class GameMap(Surface):
 
     def render(self) -> None:
         """Render the game map."""
+        # Calculate the distance the player can move in one frame.
+        # We only want to fill the area the player has moved out of, not the
+        # whole map.
+        player_step_x = self.player.max_speed * simulat.frame_delta + self.player.rect.width
+        player_step_y = self.player.max_speed * simulat.frame_delta + self.player.rect.height
+
+        # Fill the area the player has moved out of with black (color key).
+        self.character_surface.surface.fill(
+            (0, 0, 0),
+            (
+                self.player.rect.left - player_step_x,
+                self.player.rect.top - player_step_y,
+                self.player.rect.right + player_step_x,
+                self.player.rect.bottom + player_step_y
+            )
+        )
+
+        # Render the visible parts of the tile surface onto the main
+        # (game_map) surface.
         self.blit(
             self.tile_surface.surface,
             (0, 0),
+            self.camera.rect
         )
+
+        # Render the player onto the character surface.
         self.player.render()
+
+        # Render the visible parts of character surface onto the main
+        # (game_map) surface.
+        self.blit(
+            self.character_surface.surface,
+            (0, 0),
+            self.camera.rect
+        )
 
     @time_it
     def _init_tiles(self):