## PR type (check all applicable)

- [ ] `   feat   ` :sparkles: features
- [ ] `   fix    ` :bug: bugfixes
- [x] `  chore   ` :ticket: chores
- [ ] `refactor` :package: code refactoring
- [ ] `  style   ` :gem: style
- [ ] `   docs   ` :book: documentation changes
- [x] `   perf   ` :rocket: performance improvements
- [ ] `   test   ` :rotating_light: tests
- [ ] `  build   ` :construction_worker: build
- [ ] `    ci    ` :robot: continuous integration
- [ ] `  revert  ` :back: reverts

## PR description

This PR:
- Adds `GameMap.character_surface` on which characters will be rendered on.
- **Optimizes the game map rendering process.**

## Related Tickets

- related issue #
- closes #

## Instructions, Screenshots

A comparison of average (from 10 seconds of total runtime) frame rate, frame time and rendering time:
| Revision         | Framerate  | Frametime | Rendertime |
|------------------|------------|-----------|------------|
| abda27c (before) | 135.85 FPS | 7.40 ms   | 6.48 ms    |
| dca5111 (after)  | 352.96 FPS | 2.89 ms   | 2.07 ms    |

Improvement:
- Frame rate: `259.81%` more
- Render time: `313.04%` quicker